+++ plugin-build/plugin/src/test/java/com/ncorti/ktfmt/gradle/KtfmtExtensionTest.kt
@@ -84,7 +84,7 @@ class KtfmtExtensionTest {
         extension.blockIndent.set(43)
         extension.continuationIndent.set(44)
         extension.removeUnusedImports.set(false)
-        extension.manageTrailingCommas.set(true)
+        extension.trailingCommaManagementStrategy.set(TrailingCommaManagementStrategy.COMPLETE)
         extension.debuggingPrintOpsAfterFormatting.set(true)
 
         val bean = extension.toFormattingOptions()
@@ -93,7 +93,8 @@ class KtfmtExtensionTest {
         assertThat(bean.blockIndent).isEqualTo(43)
         assertThat(bean.continuationIndent).isEqualTo(44)
         assertThat(bean.removeUnusedImports).isEqualTo(false)
-        assertThat(bean.manageTrailingCommas).isEqualTo(true)
+        assertThat(bean.trailingCommaManagementStrategy)
+            .isEqualTo(TrailingCommaManagementStrategy.COMPLETE)
         assertThat(bean.debuggingPrintOpsAfterFormatting).isEqualTo(true)
     }