+++ CHANGELOG.md
@@ -5,7 +5,10 @@ Please add your entries according to this format.
 
 ## Unreleased
 - Add support for android projects with the `com.android.kotlin.multiplatform.library` plugin (#422)
+- KtFmt to 0.58
+- Deprecated `manageTrailingCommas` in favor of new `trailingCommaManagementStrategy` property
 
+##  Version 0.23.0
 - Add `ktfmt.useClassloaderIsolation` property to toggle between processIsolation and classloaderIsolation for the Gradle Worker (#427)
 - Kotlin to 2.1.21
 - Gradle to 8.14.1